feat: Add login functionality to AccountController
This commit adds the necessary code to the AccountController to handle user login. It includes the implementation of the Login action method for both GET and POST requests, as well as the Logout action method. The Login view is also added to the Views/Account folder.

The changes in this commit allow users to enter their username and log in to the application. If the username is empty, an error message is displayed. If the user is already authenticated, the navigation bar shows the user's name and a logout button. If the user is not authenticated, the navigation bar shows a login button.

Note: This commit message follows the conventional format of "feat: <description>".